requirements.txt
pyTelegramBotAPI==4.22.0
python-dotenv==1.0.1
requests==2.32.3
cachetools==5.5.0
rapidfuzz==3.9.7
